e: add new configuration key IPC_BASE_URI
According to an operation bug on ebiothon platform reported by nbard,
the /tmp mount point is regularly I/O satured which prevents the correct
working of the tmssed on 3.2.1
Since the SeD uses by default (non-configurable) a ZMQ IPC socket (Unix
socket behind the scenes) in /tmp, it was not fixable.
I added a configuration key which allows to override the default path
aka /tmp/vishnu-

SHOULD: fix all the ZMQ IPC sockets to use this key.